Cast depth delta toInt
We are ensured that the depth delta will not be larger than an Int,
as the depth isn't larger than an Int.

+++ casper/src/main/scala/coop/rchain/casper/util/ProtoUtil.scala
@@ -54,7 +54,7 @@ object ProtoUtil {
                       for {
                         updatedEstimate <- unsafeGetBlock[F](mainParentHash)
                         depthDelta      = blockNumber(updatedEstimate) - blockNumber(estimate)
-                        newDepth        = depth + depthDelta
+                        newDepth        = depth + depthDelta.toInt
                         mainChain <- if (newDepth <= 0) {
                                       (acc :+ estimate).pure[F]
                                     } else {
